Improved performance of ISTA

+++ ISTA/test_ista.h
@@ -21,14 +21,14 @@
 
 void TestIsta( uint num_rows, uint num_cols ) {
 
-    auto X = build_matrix<double>( num_rows, num_cols, &eucl_distance );
+    auto X = build_matrix<float>( num_rows, num_cols, &eucl_distance );
     auto Y = X.col(0);
-    auto W_0 = Eigen::Matrix< double, Eigen::Dynamic, 1 > ( num_rows, 1 );
+    auto W_0 = Eigen::Matrix< float, Eigen::Dynamic, 1 > ( num_rows, 1 );
     W_0.setZero();
 
-    double lambda = 1.0;
+    float lambda = 1.0;
 
-    auto ista_retval = ISTA< double >( X, Y, W_0, 1, 0.1, 0.5*lambda );
+    auto ista_retval = ISTA< float >( X, Y, W_0, 1, 0.1, 0.5*lambda );
 
     std::cout << "ISTA result:\n" << ista_retval << std::endl;
 }
@@ -42,7 +42,7 @@ void RunIstaTests() {
                   << "x" \
                   << k \
                   << "Matrix: \n" \
-                  << build_matrix<double>( k, k, &eucl_distance ) \
+                  << build_matrix<float>( k, k, &eucl_distance ) \
                   << std::endl;
 
         TestIsta( k, k );
