+++ SPAMS/test_fista.h
@@ -20,14 +20,14 @@
 
 void TestFistaFlat( uint num_rows, uint num_cols ) {
 
-    auto X = build_matrix<double>( num_rows, num_cols, &eucl_distance );
+    auto X = build_matrix<float>( num_rows, num_cols, &eucl_distance );
     auto Y = X.col(0);
-    auto W_0 = Eigen::Matrix< double, Eigen::Dynamic, 1 > ( num_rows, 1 );
+    auto W_0 = Eigen::Matrix< float, Eigen::Dynamic, 1 > ( num_rows, 1 );
     W_0.setZero();
 
-    double lambda = 1.0;
+    float lambda = 1.0;
 
-    auto spams_retval =  FistaFlat< double >( Y, X, W_0, 0.5*lambda );
+    auto spams_retval =  FistaFlat< float >( Y, X, W_0, 0.5*lambda );
 
     std::cout << "fistaFlat result:\n" << spams_retval << std::endl;
 }
@@ -41,7 +41,7 @@ void RunFistaTests() {
                   << "x" \
                   << k \
                   << "Matrix: \n" \
-                  << build_matrix<double>( k, k, &eucl_distance ) \
+                  << build_matrix<float>( k, k, &eucl_distance ) \
                   << std::endl;
 
         TestFistaFlat( k, k );
